backend/requirements.txt
# Backend Django
Django==4.2.7
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.3.1

# Filtres
django-filter==23.3

# Images
Pillow==10.1.0
django-imagekit==5.0.0

# Tâches asynchrones (optionnel pour l'instant)
# celery==5.3.4
# redis==5.0.1

# Validation
# django-ratelimit==4.1.0

# Tests et développement
colorama==0.4.6
requests==2.31.0

#pip install djangorestframework-simplejwt django-filter django-imagekit